+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Additions
 Improvements
 - Allow comparing Keybind and Combination based on the amount of modifiers they have
 - Allow using multiple textures in a StaticSpriteBatch
+- Added GenericInput support for Buttons.None
 
 Removals
 - Marked AStar.InfiniteCost as obsolete